requirements.txt
fastapi
uvicorn
nylas
python-dotenv
httpx
motor
boto3
opencv-python-headless
moviepy